refactor(notes): move editor mode from Appearance to Behavior
Editor mode (Markdown source vs Live preview) is "how you edit", not
"how the app looks", so it belongs with the other editing settings.
After the Behavior page landed (#366) it was the only editing control
left under Appearance, split from its twin default open mode that users
look for together (#351).

- Move the editor mode card (UI + handler + store subscription) from
  settings/appearance to settings/behavior, between open mode and
  delete confirmation.
- i18n (5 locales): retarget the intro-dialog hint from "Settings ->
  Appearance" to "Settings -> Behavior" (localized page name); move
  settings_page.appearance.editor_mode* keys to
  settings_page.behavior.editor_mode* so the key path matches the UI;
  extend behavior.hub_desc with "editor mode".

No schema/sync/crypto change: editorMode stays the same per-app synced
AppSettings field; only its UI location and the modal pointer move.
