+++ apps/reborn-notes/src/routes/settings/behavior/+page.svelte
@@ -14,11 +14,13 @@
     Switch,
     toast
   } from '@reborn/ui';
-  import { Eye, Trash2 } from '@lucide/svelte';
+  import { Eye, Pencil, Trash2 } from '@lucide/svelte';
   import {
     appSettings,
-    confirmBeforeDelete as confirmBeforeDeleteStore
+    confirmBeforeDelete as confirmBeforeDeleteStore,
+    editorMode as editorModeStore
   } from '$lib/stores/app-settings.store';
+  import type { EditorMode } from '@reborn/storage';
   import {
     devicePrefs,
     noteOpenMode as noteOpenModeStore,
@@ -32,6 +34,7 @@
   const isMobile = useIsMobile();
 
   let openModeValue = $state<NoteOpenMode>('preview');
+  let editorModeValue = $state<EditorMode>('markdown');
   let confirmDeleteValue = $state(true);
 
   // Split is a desktop layout (editor + preview side by side). On phones the app
@@ -55,6 +58,17 @@
     toast.success($t('settings_page.behavior.open_mode_updated'));
   }
 
+  async function updateEditorMode(value: string | undefined) {
+    if (value !== 'markdown' && value !== 'live') return;
+    try {
+      await appSettings.update('editorMode', value);
+      editorModeValue = value;
+      toast.success($t('settings_page.behavior.editor_mode_updated'));
+    } catch (err: unknown) {
+      logger.error('Failed to update editor mode', err);
+    }
+  }
+
   async function updateConfirmDelete(value: boolean) {
     try {
       await appSettings.update('confirmBeforeDelete', value);
@@ -68,6 +82,7 @@
   onMount(() => {
     const unsubscribes = [
       noteOpenModeStore.subscribe((v) => (openModeValue = v)),
+      editorModeStore.subscribe((v) => (editorModeValue = v)),
       confirmBeforeDeleteStore.subscribe((v) => (confirmDeleteValue = v))
     ];
     return () => unsubscribes.forEach((fn) => fn());
@@ -103,6 +118,36 @@
         <p class="text-xs text-muted-foreground">
           {$t('settings_page.behavior.open_mode_device_hint')}
         </p>
+      </CardContent>
+    </Card>
+
+    <!-- Editor mode (synced across devices) -->
+    <Card>
+      <CardHeader>
+        <CardTitle class="text-base flex items-center gap-2">
+          <Pencil class="h-4 w-4 text-muted-foreground" />
+          {$t('settings_page.behavior.editor_mode')}
+        </CardTitle>
+        <CardDescription>{$t('settings_page.behavior.editor_mode_desc')}</CardDescription>
+      </CardHeader>
+      <CardContent class="space-y-3">
+        <Select
+          type="single"
+          value={editorModeValue}
+          onValueChange={(value) => updateEditorMode(value)}
+        >
+          <SelectTrigger class="w-full">
+            {editorModeValue === 'live'
+              ? $t('settings_page.behavior.editor_mode_live')
+              : $t('settings_page.behavior.editor_mode_markdown')}
+          </SelectTrigger>
+          <SelectContent>
+            <SelectItem value="markdown"
+              >{$t('settings_page.behavior.editor_mode_markdown')}</SelectItem
+            >
+            <SelectItem value="live">{$t('settings_page.behavior.editor_mode_live')}</SelectItem>
+          </SelectContent>
+        </Select>
       </CardContent>
     </Card>
