+++ packages/i18n/src/translations/notes/fr.json
@@ -773,7 +773,7 @@
     "option_markdown_description": "Les marqueurs bruts comme ** # [texte](lien) restent visibles. Pour les utilisateurs qui préfèrent un contrôle total sur la syntaxe.",
     "recommended": "Recommandé",
     "hint_switch_before_icon": "Vous pouvez changer de mode à tout moment via le bouton",
-    "hint_switch_after_icon": "dans l'en-tête de la note ou dans Paramètres → Apparence.",
+    "hint_switch_after_icon": "dans l'en-tête de la note ou dans Paramètres → Comportement.",
     "continue_button": "Commencer"
   },
   "save_status": {
@@ -807,7 +807,7 @@
     },
     "behavior": {
       "title": "Comportement",
-      "hub_desc": "Mode d'ouverture des notes, confirmation de suppression",
+      "hub_desc": "Mode d'ouverture des notes, mode d'édition, confirmation de suppression",
       "open_mode": "Ouvrir les notes en",
       "open_mode_desc": "Vue dans laquelle une note existante s'ouvre lorsque vous la sélectionnez. Les nouvelles notes s'ouvrent toujours prêtes à être modifiées.",
       "open_mode_preview": "Aperçu (lecture seule)",
@@ -815,6 +815,11 @@
       "open_mode_split": "Partagé (édition et aperçu)",
       "open_mode_device_hint": "S'applique uniquement à cet appareil",
       "open_mode_updated": "Mode d'ouverture par défaut mis à jour",
+      "editor_mode": "Mode d'édition",
+      "editor_mode_desc": "Comment le Markdown est affiché pendant l'édition d'une note",
+      "editor_mode_markdown": "Markdown (afficher toute la syntaxe)",
+      "editor_mode_live": "Aperçu en direct (masquer la syntaxe hors du curseur)",
+      "editor_mode_updated": "Mode d'édition mis à jour",
       "confirm_delete": "Confirmer avant de supprimer",
       "confirm_delete_desc": "Demander avant de déplacer une note vers la Corbeille. Les notes dans la Corbeille peuvent être restaurées, mais sont définitivement supprimées après 30 jours.",
       "confirm_delete_on": "Confirmation activée",
@@ -843,12 +848,7 @@
       "image_loading_always": "Toujours charger automatiquement",
       "image_loading_never": "Ne jamais charger les images externes",
       "image_loading_updated": "Paramètre de chargement des images mis à jour",
-      "image_loading_warning": "Le chargement automatique des images peut révéler votre adresse IP à des serveurs externes",
-      "editor_mode": "Mode d'édition",
-      "editor_mode_desc": "Comment le Markdown est affiché pendant l'édition d'une note",
-      "editor_mode_markdown": "Markdown (afficher toute la syntaxe)",
-      "editor_mode_live": "Aperçu en direct (masquer la syntaxe hors du curseur)",
-      "editor_mode_updated": "Mode d'édition mis à jour"
+      "image_loading_warning": "Le chargement automatique des images peut révéler votre adresse IP à des serveurs externes"
     },
     "security": {
       "title": "Sécurité",
